+++ flecsi/exec/hpx/policy.hh
@@ -94,7 +94,7 @@ reduce_internal(Args &&... args) {
         .template delay_execution<R>(
           util::symbol<F>(), std::move(apply_delayed_prolog)));
     bound_params.set_future(ret.depend());
-    if(launch::mpi)
+    if(launch::sync)
       ret.wait();
     return ret;
   };